pyproject.toml
[project]
name = "remove-pdf-password"
version = "0.1.0"
description = "PDF 密碼移除工具"
authors = [
    {name = "gene.su"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "tkinterdnd2 (>=0.4.3,<0.5.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
